Completed Ghost Movement
I finished automating the ghost's movement.

Found a bug: the player can get stuck in the wall due to the automatic collision mask, need to set a custom collision mask manually for the main part of the player character. Waiting to do this for when we change the player sprite (assuming it's not staying a frog).